refactor: application 이벤트 DTO 패키지 위치 정리
