+++ order-service/src/main/java/com/palja/order_service/application/event/dto/request/OrderCreatedEventReq.java
@@ -1,6 +1,5 @@
-package com.palja.order_service.application.dto.event.request;
+package com.palja.order_service.application.event.dto.request;
 
-import com.palja.order_service.domain.entity.Order;
 import lombok.*;
 
 import java.util.UUID;
