+++ order-service/src/main/java/com/palja/order_service/application/saga/steps/ApplyCouponStep.java
@@ -1,7 +1,7 @@
 package com.palja.order_service.application.saga.steps;
 
-import com.palja.order_service.application.dto.event.request.CouponCancelEventReq;
-import com.palja.order_service.application.dto.event.request.CouponUseEventReq;
+import com.palja.order_service.application.event.dto.request.CouponCancelEventReq;
+import com.palja.order_service.application.event.dto.request.CouponUseEventReq;
 import com.palja.order_service.application.port.OrderEventPublisher;
 import com.palja.order_service.application.saga.SagaStep;
 import com.palja.order_service.application.saga.model.OrderSaga;
