+++ order-service/src/main/java/com/palja/order_service/infrastructure/external/kafka/publisher/OrderKafkaProducer.java
@@ -1,6 +1,6 @@
 package com.palja.order_service.infrastructure.external.kafka.publisher;
 
-import com.palja.order_service.application.dto.event.request.*;
+import com.palja.order_service.application.event.dto.request.*;
 import com.palja.order_service.application.port.OrderEventPublisher;
 import com.palja.order_service.infrastructure.external.kafka.KafkaTopics;
 import lombok.RequiredArgsConstructor;
